renamed KTrajectorySeq to KTrajectoryPulseq
